Refactored project structure for consistency
- Moved landing components from app/components/landing/ to app/landing/components/
- Updated all import paths to reflect new file locations
