+++ app/landing/components/Landing.tsx
@@ -1,5 +1,5 @@
-import Header from "../layout/Header";
-import Footer from "../layout/Footer";
+import Header from "../../components/layout/Header";
+import Footer from "../../components/layout/Footer";
 import Countdown from "./countdowns/top_countdown/TopCountdownSection";
 import Hero from "./hero/Hero";
 import AboutConferenceSection from "./about_conference/AboutConferenceSection";
