+++ app/landing/components/about_conference/AboutConferenceSection.tsx
