+++ app/landing/components/faq/FAQSection.tsx
